+++ src/Beef.Core/Mapper/PropertyMapperCustomBase.cs
@@ -64,6 +64,11 @@ namespace Beef.Mapper
         public Type SrcePropertyType => typeof(TSrceProperty);
 
         /// <summary>
+        /// Gets the destination property <see cref="Type"/>.
+        /// </summary>
+        public virtual Type DestPropertyType => null;
+
+        /// <summary>
         /// Gets the destination property name.
         /// </summary>
         public string DestPropertyName { get; protected set; }