+++ src/Beef.Core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Represents the **NuGet** versions.
 
+## v2.1.18
+- *Added:* Support for `IConvertible` added to `ReferenceDataBase` to enable usage of `Convert.ChangeType`.
+- *Added:* `PropertyMapper` updated to use `Convert.ChangeType` as last resort property value mapping.
+
 ## v2.1.17
 - *Added:* New `Beef.Event.Subscribe` namespace added to enable the base capabilities for the subscription of events.